Unified the way of updating hosts file for OSX and *nix
